Fixed the API url to be relative and cleaned up the code.
The API server side url request was a specific url, so it won't work on a different server. Fixed it so that it it's relative to the current directory. This can be adjusted later when the front-end code is merged with the server side code, but should work on any server now. Also cleaned up the code.