Admin can select start location and upload image of a car
Admin can now select the start location of their car from a drop down list. They can now chose an image for the car which will be uploaded to images/cars/id_of_car.extension. The extension is stored in the cars table under imageExtension. Will add an easy to use api that returns this.